Will throw error in messages if no OXR ID provided
If you try to use the OXR backend but do not provide an ID, a warning
should be thrown.

Right now the error handling ain't great.  It'll leave a "Please sign
up for an OpenExchangeRates App ID!" message in the *Messages* buffer,
but I'd like it to bubble up to the top and not be covered up by other
warnings...